refactor/create-auth-package
- 인증 관련 기능을 `auth` 패키지로 이동시켰습니다.
- "이메일, 비밀번호를 통한 유저 정보 가져오기" 기능은 `User` 도메인에 관련된 기능이기 때문에 `UserService` 로 이동시켰습니다.
