+++ src/test/java/com/airjnc/user/controller/UserControllerTest.java
@@ -21,10 +21,10 @@ import com.airjnc.user.dto.response.UserResp;
 import com.airjnc.user.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.testutil.annotation.IntegrationTest;
-import com.testutil.fixture.UserCreateReqFixture;
-import com.testutil.fixture.UserInquiryEmailReqDTOFixture;
-import com.testutil.fixture.UserInquiryEmailResDTOFixture;
-import com.testutil.fixture.UserRespFixture;
+import com.testutil.fixture.user.UserCreateReqFixture;
+import com.testutil.fixture.user.UserInquiryEmailReqDTOFixture;
+import com.testutil.fixture.user.UserInquiryEmailResDTOFixture;
+import com.testutil.fixture.user.UserRespFixture;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.Test;
