--- src/test/java/com/airjnc/user/service/AuthServiceTest.java
@@ -1,63 +1,0 @@
-package com.airjnc.user.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
-import static org.mockito.Mockito.times;
-import com.airjnc.common.service.StateService;
-import com.airjnc.common.util.enumerate.SessionKey;
-import com.airjnc.user.dao.UserRepository;
-import com.airjnc.user.domain.UserEntity;
-import com.airjnc.user.dto.UserWhereDto;
-import com.airjnc.user.dto.request.AuthLogInReq;
-import com.airjnc.user.dto.response.UserResp;
-import com.airjnc.user.util.UserModelMapper;
-import com.testutil.annotation.UnitTest;
-import com.testutil.fixture.AuthLogInReqFixture;
-import com.testutil.fixture.UserRespFixture;
-import com.testutil.testdata.TestUser;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-@UnitTest
-class AuthServiceTest {
-
-  @Mock
-  UserModelMapper userModelMapper;
-
-  @Mock
-  UserRepository userRepository;
-
-  @Mock
-  UserCheckService userCheckService;
-
-  @Mock
-  StateService stateService;
-
-  @InjectMocks
-  AuthService authService;
-
-  @Test
-  void userShouldBeLoggedIn() {
-    //given
-    AuthLogInReq authLogInReq = AuthLogInReqFixture.getBuilder().build();
-    UserEntity userEntity = TestUser.getBuilder().build();
-    UserResp userResp = UserRespFixture.getBuilder().build();
-
-    given(userRepository.findByWhere(any(UserWhereDto.class))).willReturn(userEntity);
-    given(userModelMapper.userEntityToUserResp(userEntity)).willReturn(userResp);
-    //when
-    UserResp result = authService.logIn(authLogInReq);
-    //then
-    then(userRepository).should(times(1)).findByWhere(any(UserWhereDto.class));
-    then(userCheckService).should(times(1)).passwordShouldBeMatch(authLogInReq.getPassword(), userEntity.getPassword());
-    then(stateService).should(times(1)).create(SessionKey.USER, userEntity.getId());
-    then(userModelMapper).should(times(1)).userEntityToUserResp(userEntity);
-    assertThat(result).isSameAs(userResp);
-  }
-}
