+++ src/test/java/com/testutil/fixture/auth/AuthLogInReqFixture.java
@@ -1,6 +1,6 @@
-package com.testutil.fixture;
+package com.testutil.fixture.auth;
 
-import com.airjnc.user.dto.request.AuthLogInReq;
+import com.airjnc.auth.dto.request.AuthLogInReq;
 
 public class AuthLogInReqFixture {
