+++ src/test/java/com/testutil/fixture/user/UserRespFixture.java
@@ -1,4 +1,4 @@
-package com.testutil.fixture;
+package com.testutil.fixture.user;
 
 import com.airjnc.user.dto.response.UserResp;
 import com.testutil.testdata.TestUser;
